refactored SoldItem components into their own directory
